Implemented eq1.subs(eq2) (#11)
* rebasing master

* Implemented eq1.subs(eq2)

Sometime it is useful to take one equation and substitute it into another one. For example:

```
eq1 = Equation(L1, R1)
eq2 = Equation(L2, R2)
eq1.subs(eq2)
```

The above code will look for L2 inside L1 and R1: if matches are found, they will be replaced
with R2. This feature is very important for this class to be used "naturally" and quickly.
The obvious alternative would be to execute:

```
eq1.subs(eq2.lhs, eq2.rhs)
```

which is fine if we are only doing one substitution once in a while, but become time consuming
when dealing with multiple equations and multiple substitutions.

Even though the best approach would be to tweak the `Basic` class to make it recognizes `Equation`,
at this point of development it is not worth the effort. For the time being it is better to get
`Equation` overrides the `subs` method in order to implement the new behaviour. If and when this
module will be merged into sympy, then editing `Basic.subs` will be trivial.
This means that substituting an equation into a symbolic expression will fail.

* removed code that slipped in from a previous branch

* added error message and improved tests