Fixed latex printer (#16)
The `_latex` method was implemented in a wrong way, which didn't allowed the use of a custom printer.
The capability to use custom printer is essential. Think for example to an equation containing multiple
undefined applied functions, `f(x, y, z), g(x, y, z), ...`. We quickly run out of display
space on a Jupyter Notebook cell because `(x, y, z)` is rendered multiple times.

According to the official documentation:
https://docs.sympy.org/latest/modules/printing.html
it should be `def _latex(self, printer)`.

Implementing it in this way allows the use of a custom printer. Tests have been added to assert this
behavior.